requirements.txt
selenium
pytest
pytest-xdist
pytest-html
requests
webdriver-manager
pyyaml
allure-pytest